+++ iModelCore/ECDb/CHANGES.md
@@ -5,7 +5,15 @@ This document including important changes to syntax or file format.
 | Module  | Version   |
 | ------- | --------- |
 | Profile | `4.0.0.5` |
-| ECSQL   | `2.0.2.0` |
+| ECSQL   | `2.0.2.1` |
+
+## ## `05/12/2025`: Added PRAGMA `validate_ecsql_inserts` to validate navigation property inserts
+* ECSql version change `2.0.2.0` -> `2.0.2.1`.
+* Add `PRAGMA validate_ecsql_inserts` allow to enable validation of RelECClassId when inserting a navigation property of a class.
+* This config defaults to false.
+* The navigation value binders have also been updated to handle the validations when the config is enabled.
+* The validation only covers the navigation properties in ECSql insert statements.
+
 
 ## ## `01/29/2025`: Made schema names optional for table valued functions
 * ECSql version change `2.0.1.1` -> `2.0.2.0`.